feat: Add `xAlign`/`yAlign` props
Add new CSS variables to manage the new align-items and justify-content values, as well as flex-direction.

Add a new useAlign() hook to correctly handle switching between align-items/justify-content based on the flex-direction value.

With this change it is possible to use normalized `xAlign` (for horizontal alignment) and `yAlign` (for horizontal alignment) props.